+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Changed
+
+    - [config] changed default user in `settings.yaml` to empty
+
+### Fixed
+
+    - [cmd/new] fix: explicitly empty labels not handled correctly (GitHub Issue #90)
+
 ## [v0.6.1] - 2026-01-05
 
 ### Added